[SCTP] Refactor SCTP socket creation to use address family from sa_list (#3344)
Refactored the SCTP socket creation logic to dynamically select
the address family based on the provided address list.

A new function, `create_sctp_socket_from_addr_list`, was introduced
to check for the presence of an IPv6 address in the address list and
create an SCTP socket accordingly.

If an IPv6 address is found, it uses AF_INET6; otherwise, it defaults
to AF_INET. This change was applied to both the `ogs_sctp_server` and
`ogs_sctp_client` functions.